Move classes to lsst.ts.salpytools
