+++ python/lsst/ts/__init__.py
@@ -1,0 +1,2 @@
+import pkgutil, lsstimport
+__path__ = pkgutil.extend_path(__path__, __name__)
