+++ python/lsst/ts/salpytools/state_transition_exception.py
