+++ python/lsst/ts/salpytools/utils.py
